requirements.txt
djangorestframework
django
redis
celery
tk
selenium
requests